Fixed dupPageCheck(), Added csvCheck()
dupPageCheck() will now add " (n)" to the end of the name where n is a
number that counts up if the name still exists.
csvCheck() will check the csv for common errors and call them out before
the pages are created. This will save time by finding errors at the
start of running.